Abstracts reducer logic into helpers
