+++ app/javascript/react/screens/App/Overview/screens/MappingWizard/components/MappingWizardClustersStep/helpers.js
@@ -1,0 +1,20 @@
+export const getHostsByClusterID = hostsQueryFulfilledAction => {
+  const {
+    meta: { hostIDsByClusterID },
+    payload: { data }
+  } = hostsQueryFulfilledAction;
+  const hostsByID = data.results.reduce(
+    (newObject, host) => ({
+      ...newObject,
+      [host.id]: host
+    }),
+    {}
+  );
+  return Object.keys(hostIDsByClusterID).reduce(
+    (newObject, clusterID) => ({
+      ...newObject,
+      [clusterID]: hostIDsByClusterID[clusterID].map(hostID => hostsByID[hostID])
+    }),
+    {}
+  );
+};
