+++ app/javascript/react/screens/App/Plan/PlanReducer.js
@@ -1,8 +1,5 @@
 import Immutable from 'seamless-immutable';
-import numeral from 'numeral';
 import commonUtilitiesHelper from '../common/commonUtilitiesHelper';
-import getMostRecentVMTasksFromRequests from '../Overview/components/Migrations/helpers/getMostRecentVMTasksFromRequests';
-import { IsoElapsedTime } from '../../../../components/dates/IsoElapsedTime';
 
 import {
   FETCH_V2V_PLAN,
@@ -13,8 +10,6 @@ import {
   DOWNLOAD_LOG_CLICKED,
   DOWNLOAD_LOG_COMPLETED,
   FETCH_V2V_ANSIBLE_PLAYBOOK_TEMPLATE,
-  V2V_MIGRATION_STATUS_MESSAGES,
-  STATUS_MESSAGE_KEYS,
   FETCH_V2V_ORCHESTRATION_STACK,
   REMOVE_TASKS_SELECTED_FOR_CANCELLATION,
   UPDATE_TASKS_SELECTED_FOR_CANCELLATION,
@@ -22,6 +17,15 @@ import {
   ADD_TASKS_TO_MARKED_FOR_CANCELLATION,
   ADD_TASK_TO_NOTIFICATION_SENT_LIST
 } from './PlanConstants';
+import {
+  excludeDownloadDoneTaskId,
+  includeDownloadInProgressTaskId,
+  getFailedMigrations,
+  getSuccessfulMigrations,
+  allVMTasksForRequestOfPlan,
+  incompleteCancellationTasks,
+  removeCancelledTasksFromSelection
+} from './helpers';
 
 export const initialState = Immutable({
   isFetchingPlanRequest: false,
@@ -53,117 +57,6 @@ export const initialState = Immutable({
   successfulMigrations: [],
   notificationsSentList: []
 });
-
-const excludeDownloadDoneTaskId = (allDownloadLogInProgressTaskIds, taskId) =>
-  allDownloadLogInProgressTaskIds.filter(element => element !== taskId);
-
-const includeDownloadInProgressTaskId = (allDownloadLogInProgressTaskIds, taskId) =>
-  allDownloadLogInProgressTaskIds ? allDownloadLogInProgressTaskIds.concat(taskId) : [taskId];
-
-const removeCancelledTasksFromSelection = (allSelectedTasksForCancel, alreadyCancelledTasks) =>
-  allSelectedTasksForCancel.filter(selectedTask =>
-    alreadyCancelledTasks.every(cancelledTask => selectedTask.id !== cancelledTask.id)
-  );
-
-const taskCompletedSuccessfully = task => task.state === 'finished' && task.status === 'Ok';
-
-const taskCompletedUnsuccessfully = task => task.state === 'finished' && task.status !== 'Ok';
-
-const getFailedMigrations = vmTasks => vmTasks.filter(task => taskCompletedUnsuccessfully(task));
-
-const getSuccessfulMigrations = vmTasks => vmTasks.filter(task => taskCompletedSuccessfully(task));
-
-const processVMTasks = vmTasks => {
-  const tasks = [];
-  vmTasks.forEach(task => {
-    const taskDetails = {
-      id: task.id,
-      message: V2V_MIGRATION_STATUS_MESSAGES[task.message] || __('VM migrations stalled'),
-      transformation_host_name: task.options && task.options.transformation_host_name,
-      delivered_on: new Date(task.options.delivered_on),
-      updated_on: new Date(task.updated_on),
-      completed:
-        task.message === STATUS_MESSAGE_KEYS.VM_MIGRATIONS_COMPLETED ||
-        task.message === STATUS_MESSAGE_KEYS.VM_MIGRATIONS_FAILED ||
-        task.message === STATUS_MESSAGE_KEYS.FAILED ||
-        (!V2V_MIGRATION_STATUS_MESSAGES[task.message] && task.state === 'finished'),
-      state: task.state,
-      status: task.status,
-      options: {},
-      cancel_requested: task.options.cancel_requested
-    };
-
-    if (task.options.playbooks) {
-      taskDetails.options.prePlaybookRunning =
-        task.options.playbooks.pre && task.options.playbooks.pre.job_state === 'active';
-      taskDetails.options.postPlaybookRunning =
-        task.options.playbooks.post && task.options.playbooks.post.job_state === 'active';
-      taskDetails.options.prePlaybookComplete =
-        task.options.playbooks.pre && task.options.playbooks.pre.job_state === 'finished';
-      taskDetails.options.postPlaybookComplete =
-        task.options.playbooks.post && task.options.playbooks.post.job_state === 'finished';
-      taskDetails.options.playbooks = task.options.playbooks;
-    }
-
-    taskDetails.options.progress = task.options.progress;
-    taskDetails.options.virtv2v_wrapper = task.options.virtv2v_wrapper;
-
-    if (!task.diskSpaceCompletedGb) {
-      taskDetails.diskSpaceCompletedGb = '0';
-    }
-
-    if (!task.percentComplete) {
-      taskDetails.percentComplete = 0;
-    }
-
-    if (!task.totalDiskSpaceGb) {
-      taskDetails.totalDiskSpaceGb = '100%';
-    }
-
-    const grepVMName = task.description.match(/\[(.*?)\]/);
-
-    if (grepVMName) {
-      [taskDetails.descriptionPrefix, taskDetails.vmName] = grepVMName;
-    }
-
-    const startDateTime = taskDetails.delivered_on;
-    const lastUpdateDateTime = taskDetails.updated_on;
-    taskDetails.startDateTime = startDateTime;
-    taskDetails.lastUpdateDateTime = lastUpdateDateTime;
-
-    taskDetails.elapsedTime = IsoElapsedTime(new Date(startDateTime), new Date(lastUpdateDateTime));
-
-    if (taskDetails.completed) {
-      taskDetails.completedSuccessfully =
-        (task.options.progress && task.options.progress.current_description === 'Virtual machine migrated') ||
-        (task.options.progress && task.options.progress.current_description === 'Mark source as migrated');
-    }
-    if (task.options && task.options.virtv2v_disks && task.options.virtv2v_disks.length) {
-      taskDetails.totalDiskSpace = task.options.virtv2v_disks.reduce((a, b) => a + b.size, 0);
-      taskDetails.totalDiskSpaceGb = numeral(taskDetails.totalDiskSpace).format('0.00b');
-
-      const percentComplete =
-        task.options.virtv2v_disks.reduce((a, b) => a + b.percent, 0) / (100 * task.options.virtv2v_disks.length);
-
-      taskDetails.diskSpaceCompleted = percentComplete * taskDetails.totalDiskSpace;
-      taskDetails.diskSpaceCompletedGb = numeral(taskDetails.diskSpaceCompleted).format('0.00b');
-      taskDetails.percentComplete = Math.round(percentComplete * 100);
-    }
-    tasks.push(taskDetails);
-  });
-  return tasks;
-};
-
-const allVMTasksForRequestOfPlan = (requestWithTasks, actions) => {
-  const tasksOfPlan = getMostRecentVMTasksFromRequests(requestWithTasks, actions);
-  return processVMTasks(tasksOfPlan);
-};
-
-const incompleteCancellationTasks = (requestWithTasks, actions, tasksForCancel) => {
-  const tasksOfPlan = getMostRecentVMTasksFromRequests(requestWithTasks, actions);
-  const completedTasksOfPlan = tasksOfPlan.filter(task => task.state === 'finished');
-  return removeCancelledTasksFromSelection(tasksForCancel, completedTasksOfPlan);
-};
 
 export default (state = initialState, action) => {
   switch (action.type) {
